+++ source/extensions/omni.isaac.synthetic_utils/omni/isaac/synthetic_utils/writers/__init__.py
@@ -10,4 +10,3 @@
 from .base import BaseWriter
 from .numpy import NumpyWriter
 from .kitti import KittiWriter
-from .dope import DOPEWriter